% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_std.R
\name{batch_get_gold}
\alias{batch_get_gold}
\title{Extract text from a set of PDFs with embedded text.}
\usage{
batch_get_gold(files, gold_path)
}
\arguments{
\item{gold_path}{Path to which gold-standard TXTs are written}

\item{...}{List/vector of paths to PDFs to be processed with \link{get_gold}}
}
\value{
A summary (data.frame) of the files
}
\description{
Extract text from a set of PDFs with embedded text.
}
\examples{
# res <- batch_get_gold(list.files("PDFs", pattern = "pdf"), "GOLDs")
}
\seealso{
\link{get_gold}
}

