% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrade.R
\name{batch_simulate_degrade_set}
\alias{batch_simulate_degrade_set}
\title{Simulate degraded PDFs from a set of input PDFs}
\usage{
batch_simulate_degrade_set(files)
}
\arguments{
\item{files}{List or vector of PDFs to be degraded}
}
\description{
Simulate degraded PDFs from a set of input PDFs
}

