% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_std.R
\name{check_embed}
\alias{check_embed}
\title{Check if text embed is not from OCR}
\usage{
check_embed(file)
}
\arguments{
\item{file}{Path to a PDF to check for embedding source}
}
\value{
Logical: TRUE if good embed, FALSE if from OCR
}
\description{
Some PDFs have an embedded text layer that is derived from OCR by the scanner
or other equipment that produced the PDF. Such documents will likely have
OCR artifacts that will contaminate the 'gold standard' that is needed for
error correction. The gold standard texts should only come from PDFs derived
directly from the original document (e.g., .docx).
}
\examples{
# res <- summarize_gold("test.pdf", text)
}
\seealso{
\code{pdftools::pdf_info}
}

