% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dirs}
\alias{create_dirs}
\title{Create directories for \code{ocrerrs}}
\usage{
create_dirs()
}
\value{
None. Message printed.
}
\description{
\code{ocrerrs} package expects particular directories will exist if run
in full. This creates those expected directories in the working directory.
\describe{
  \item{GOLDs}{Contains .txt and .rda of gold-standard text}
  \item{SPLITs}{Contains one subdir for each PDF, and subdirs contain
  one file per page of the PDF}
  \item{IMGs}{Contains one subdir for each PDF, and subdirs contain one .tif
  or .png per page of the PDF}
  \item{TXTs}{Contains one subdir for each PDF, and subdirs contain one .txt
  from OCR of each page of the PDF}
  \item{BADs}{Contains .txt of the PDF, from concatenating the files in
  TXTs/<pdfname>}
  \item{RES}{Contains results files from OCR error-finding}
}
}
\examples{
# create_dirs()
}

