% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrade.R
\name{degrade_blur}
\alias{degrade_blur}
\title{Degrade PDF quality by simulating blurred text}
\usage{
degrade_blur(file, blur)
}
\arguments{
\item{file}{Path to a text-embedded PDF to degrade}

\item{blur}{Mean and s.d. blur in format 'm'x's', e.g., 2x2}
}
\value{
A data.frame with the in and out file names and exit status
}
\description{
Low-resolution image-based PDFs are particularly susceptible to OCR errors.
By simulating blurred text from text-embedded PDFs we may find patterns of
OCR errors that improve post-OCR correction.
}
\examples{
# degrade_blur("good.pdf", 2x2)
}
\seealso{
\link{degrade_density} \link{degrade_fax} \link{degrade_complex} \link{degrade_rotate}
}

