% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrade.R
\name{degrade_density}
\alias{degrade_density}
\title{Degrade PDF quality by manipulating pixel density}
\usage{
degrade_density(file, density)
}
\arguments{
\item{file}{Path to a text-embedded PDF to degrade}

\item{density}{Pixel density (dpi) to render; integer}
}
\value{
A data.frame with the in and out file names and exit status
}
\description{
Low-resolution image-based PDFs are particularly susceptible to OCR errors.
By simulating low-res PDFs from text-embedded PDFs we may find patterns of
OCR errors that improve post-OCR correction.
}
\examples{
# degrade_density("PDFs/good.pdf", 150)
}
\seealso{
\link{degrade_rotate} \link{degrade_fax} \link{degrade_complex} \link{degrade_blur}
}

