% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrade.R
\name{degrade_fax}
\alias{degrade_fax}
\title{Degrade PDF quality by simulating a fax}
\usage{
degrade_fax(file)
}
\arguments{
\item{file}{Path to a text-embedded PDF to degrade}
}
\value{
A data.frame with the in and out file names and exit status
}
\description{
Low-resolution image-based PDFs are particularly susceptible to OCR errors.
By simulating a fax from text-embedded PDFs we may find patterns of
OCR errors that improve post-OCR correction.
}
\examples{
# degrade_fax("good.pdf")
}
\seealso{
\link{degrade_density} \link{degrade_rotate} \link{degrade_complex} \link{degrade_blur}
}

