% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrade.R
\name{degrade_pages}
\alias{degrade_pages}
\title{Wrap degrade functions of split PDF files}
\usage{
degrade_pages(pdf, pages, deg_fn, ...)
}
\arguments{
\item{pdf}{Path to the PDF that was split and is to be OCR'd}
}
\value{
A data.frame with infile, outfile, and exit status
}
\description{
Wrap degrade functions of split PDF files
}
\examples{
# TBD
}

