% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrade.R
\name{degrade_rotate}
\alias{degrade_rotate}
\title{Degrade PDF quality by simulating page rotation}
\usage{
degrade_rotate(file, rotate)
}
\arguments{
\item{file}{Path to a text-embedded PDF to degrade}

\item{rotate}{Degrees to rotate the pdf; usual range 0.5 to 5; float}
}
\value{
A data.frame with the in and out file names and exit status
}
\description{
Low-resolution image-based PDFs are particularly susceptible to OCR errors.
By simulating page rotation from text-embedded PDFs we may find patterns of
OCR errors that improve post-OCR correction.
}
\examples{
# degrade_rotate("good.pdf", 0.5)
}
\seealso{
\link{degrade_density} \link{degrade_fax} \link{degrade_complex} \link{degrade_blur}
}

