% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_errors.R
\name{find_errors}
\alias{find_errors}
\title{Find errors from OCR by comparing to gold standard}
\usage{
find_errors(bad, gold)
}
\arguments{
\item{bad}{A character vector of the text from OCR}

\item{gold}{A character vector of the embedded text}
}
\description{
Find errors from OCR by comparing to gold standard
}

