% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_errors.R
\name{find_min_dists}
\alias{find_min_dists}
\title{Find the minimum string edit for each bad word}
\usage{
find_min_dists(mat)
}
\arguments{
\item{mat}{A matrix of edit distances}

\item{gold}{A character vector of the embedded text}
}
\description{
Find the minimum string edit for each bad word
}

