% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{get_bg_1grams}
\alias{get_bg_1grams}
\title{Get ngrams and counts for bad and gold strings}
\usage{
get_bg_1grams(bad, gold, n = 1)
}
\arguments{
\item{bad}{Text from OCR of same document as gold}

\item{gold}{Text extracted from embedded text layer of PDF}

\item{n}{Number of n-grams to find}
}
\description{
Get ngrams and counts for bad and gold strings
}

