% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_errors.R
\name{get_delta_words}
\alias{get_delta_words}
\title{Get words with difference frequencies between bad and gold texts}
\usage{
get_delta_words(df)
}
\arguments{
\item{df}{A data.frame of bad and gold ngrams with frequencies each}
}
\description{
Get words with difference frequencies between bad and gold texts
}

