% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_errors.R
\name{get_dist_mat}
\alias{get_dist_mat}
\title{Return a matrix of optimal string alignment distances for bad, gold words}
\usage{
get_dist_mat(bad, good)
}
\arguments{
\item{bad}{A character vector of the text from OCR}

\item{gold}{A character vector of the embedded text}
}
\description{
Return a matrix of optimal string alignment distances for bad, gold words
}

