% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_std.R
\name{get_embed_pages}
\alias{get_embed_pages}
\title{Return a vector of pages with embedded text}
\usage{
get_embed_pages(txt)
}
\arguments{
\item{txt}{An object (list of pages) from \code{pdf_tools::pdf_text}}
}
\value{
A vector of pages with an embedded text layer
}
\description{
A PDF with a text layer may also have image-only pages. To find errors, we
need to OCR only the (degraded) pages with embedded text because finding
differences between the BAD and GOLD versions of the text depends on counts
of each word...if OCR picks up words on pages that \code{pdf_text} cannot
see because they are in am image, then GOLD != BAD because of different
inputs.
}
\examples{
# to be added
}

