% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_base}
\alias{get_file_base}
\title{Return the base name of a file}
\usage{
get_file_base(file)
}
\arguments{
\item{file}{Path to the file}
}
\value{
The base name of the file, without directories or file suffix
}
\description{
More detailed description of the function
}
\examples{
get_file_base("data/test.pdf")
# test
}

