% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_std.R
\name{get_gold}
\alias{get_gold}
\title{Extract text from a PDF with embedded text.}
\usage{
get_gold(file, write = FALSE, save = TRUE)
}
\arguments{
\item{file}{Path to the PDF to be processed}

\item{write}{Whether to write the text to file [FALSE]}

\item{save}{Whether to save the text as a .rda [TRUE]}
}
\value{
List of pages with text layer if layer not from OCR; else NULL
}
\description{
Uses \code{pdftools::pdf_text} to get the text layer from PDF `file`, which
is used as the 'gold standard' against which OCR'd versions are compared.
Checks that the text layer is distilled from the original document rather
than a text layer from OCR, e.g., a scanner that OCRs.
}
\examples{
# res <- get_gold("test.pdf", "GOLDs")
}

