% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{get_ngrams}
\alias{get_ngrams}
\title{Get a set of n-grams from text}
\usage{
get_ngrams(txt, from = 1, to = 5)
}
\arguments{
\item{txt}{Character string from which n-grams are extracted}

\item{from}{The length of smallest n-gram [1]}

\item{to}{The length of largest n-gram [5]}
}
\value{
A data.frame of the ngrams and their freq in txt
}
\description{
Get a set of n-grams from text
}
\examples{
# res <- get_ngrams(text, from = 1, to = 3)
}

