% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{get_POS}
\alias{get_POS}
\title{Return a table of parts of speech}
\usage{
get_POS(txt)
}
\arguments{
\item{txt}{A character representation of a text}
}
\value{
A table of word, POS, and frequency of the combination
}
\description{
Knowing the distribution of a word's part of speech (POS) can be an
informative characteristic for correcting errors. Note that this can be
computationally expensive.
}
\examples{
# get_POS(test_txt)
}

