% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_errors.R
\name{label_delta_words}
\alias{label_delta_words}
\title{Label words as correct or errors}
\usage{
label_delta_words(df)
}
\arguments{
\item{df}{A data.frame of bad and gold ngrams with frequencies each}
}
\description{
Label words as correct or errors
}

