% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_gold_path}
\alias{make_gold_path}
\title{Create a path to which 'gold standard' results are written}
\usage{
make_gold_path(file, ext)
}
\arguments{
\item{file}{Path to the PDF to be processed}

\item{ext}{File extension of file to be written}
}
\value{
The compiled path (character)
}
\description{
Create a path to which 'gold standard' results are written
}
\examples{
make_gold_path("PDFs/test.pdf", "txt")
# GOLDs/test.txt
}

