% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_text}
\alias{normalize_text}
\title{Clean EOL characters from}
\usage{
normalize_text(txt)
}
\description{
Removes single EOL sequential characters so that sentences are not split but
paragraphs remain separated.
}
\examples{
# create_dirs()
}

