% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr.R
\name{ocr_pages}
\alias{ocr_pages}
\title{Wrap optical character recognition around a set of files}
\usage{
ocr_pages(pdf, pages, ext = "png", errfile)
}
\arguments{
\item{pdf}{Path to the PDF file to be OCR'd}

\item{pages}{A vector of pages with embedded text in gold standard}

\item{ext}{The file extension to be found, either png or tif [png]}

\item{errfile}{The file to which Tesseract STDERR is written}
}
\value{
A data.frame with
}
\description{
Uses Tesseract (which must be installed and on the $PATH) to perform optical
character recognition (OCR) on a file.
}
\examples{
# res <- tess_ocr("test.pdf")
}
\seealso{
\link{extract_text}
}

