% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_std.R
\name{save_gold_text}
\alias{save_gold_text}
\title{Save the extracted text as a .rda}
\usage{
save_gold_text(file, text)
}
\arguments{
\item{file}{Path to the PDF to be processed}

\item{text}{Text extracted using \code{pdftools::pdf_text}}

\item{gold_path}{Path to which gold-standard TXT is written}
}
\value{
None
}
\description{
Save the extracted text as a .rda
}
\examples{
# save_gold_text("test.pdf", "GOLDs", text)
}

