% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrade.R
\name{simulate_degrade_set}
\alias{simulate_degrade_set}
\title{Simulate degraded PDFs from an input PDF}
\usage{
simulate_degrade_set(file)
}
\arguments{
\item{file}{Path to the PDF to be degraded}
}
\description{
Simulate degraded PDFs from an input PDF
}

