% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_pdf}
\alias{split_pdf}
\title{Split a PDF into multiple pages}
\usage{
split_pdf(file)
}
\arguments{
\item{file}{Path to the PDF to be split}
}
\value{
None
}
\description{
Uses \href{https://www.pdflabs.com/docs/pdftk-man-page/}{PDFtk}, which must
be on the $PATH, to split the PDF.
}
\examples{
# split_pdf("PDFs/test.pdf")
}

