% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_std.R
\name{summarize_gold}
\alias{summarize_gold}
\title{Summarize the text from a gold-standard PDF}
\usage{
summarize_gold(fil, txt)
}
\arguments{
\item{fil}{Name of the PDF from which text was extracted}

\item{txt}{List from \code{pdftools::pdf_text} with text from PDF}
}
\value{
A data.frame with # pages, # lines, and # words
}
\description{
More detailed description of the function
}
\examples{
# res <- summarize_gold("test.pdf", text)
}

