% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr.R
\name{tess_ocr}
\alias{tess_ocr}
\title{Perform optical character recognition with tesseract}
\usage{
tess_ocr(file, errfile)
}
\arguments{
\item{file}{Path to file to be OCR'd}

\item{errfile}{The file to which Tesseract STDERR is written}
}
\value{
The path to the OCR'd text file
}
\description{
Uses Tesseract (which must be installed and on the $PATH) to perform optical
character recognition (OCR) on a file.
}
\examples{
# res <- tess_ocr("test.pdf")
}
\seealso{
\link{extract_text}
}

