% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_std.R
\name{write_gold_text}
\alias{write_gold_text}
\title{Write the extracted text to file}
\usage{
write_gold_text(file, text)
}
\arguments{
\item{file}{Path to the PDF to be processed}

\item{text}{Text extracted using \code{pdftools::pdf_text}}

\item{gold_path}{Path to which gold-standard TXT is written}
}
\value{
None
}
\description{
Write the extracted text to file
}
\examples{
# write_gold_text("test.pdf", "GOLDs", text)
}

